Move JsonUtils to utilities package
