+++ bundles/org.openhab.binding.icloud/src/main/java/org/openhab/binding/icloud/internal/handler/ICloudAccountBridgeHandler.java
@@ -27,11 +27,11 @@ import org.eclipse.jdt.annotation.Nullable;
 import org.openhab.binding.icloud.internal.ICloudApiResponseException;
 import org.openhab.binding.icloud.internal.ICloudDeviceInformationListener;
 import org.openhab.binding.icloud.internal.ICloudService;
-import org.openhab.binding.icloud.internal.JsonUtils;
 import org.openhab.binding.icloud.internal.RetryException;
 import org.openhab.binding.icloud.internal.configuration.ICloudAccountThingConfiguration;
 import org.openhab.binding.icloud.internal.handler.dto.json.response.ICloudAccountDataResponse;
 import org.openhab.binding.icloud.internal.handler.dto.json.response.ICloudDeviceInformation;
+import org.openhab.binding.icloud.internal.utilities.JsonUtils;
 import org.openhab.core.cache.ExpiringCache;
 import org.openhab.core.storage.Storage;
 import org.openhab.core.thing.Bridge;
