+++ bundles/org.openhab.binding.icloud/src/test/java/org/openhab/binding/icloud/TestICloud.java
@@ -26,8 +26,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.openhab.binding.icloud.internal.ICloudApiResponseException;
 import org.openhab.binding.icloud.internal.ICloudService;
-import org.openhab.binding.icloud.internal.JsonUtils;
 import org.openhab.binding.icloud.internal.handler.dto.json.response.ICloudAccountDataResponse;
+import org.openhab.binding.icloud.internal.utilities.JsonUtils;
 import org.openhab.core.storage.json.internal.JsonStorage;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
